fix: security + correctness audit — 8 critical, 7 medium bugs fixed
Critical:
- string_enc.h: ENC() now produces XOR-encrypted byte array at compile time
  USE() decrypts at runtime (was no-op passthrough — plaintext in binary)
- crypto.c: HMAC-SHA256 replaces trivial XOR MAC (was forgeable)
- crypto.c: /dev/urandom replaces LCG PRNG (was predictable)
- crypto.c: key XOR-obfuscated at compile time (was plaintext hex)
- socket_server.c: MSG_CHALLENGE sends proper HMAC challenge (was echo-back)

Medium:
- ipc_protocol.h: foreground_pid uint8_t->uint16_t (PIDs >255 truncated)
- daemon.h: removed duplicate msg_type_t (conflicted with ipc_protocol.h)
- socket_server.c: send_response() handles partial writes (was single write)
- socket_server.c: chmod 0666->0660 (world-writable socket)
- thermal_core.c: uint32_t freq written as uint32 not int (>2.147GHz wrap)
- battery_monitor.c: abs()->labs() (INT_MIN undefined behavior)
- sysfs_monitor.c: uint32_t cast checks for -1 before cast

Low:
- profile_engine.c: (void)json suppresses unused param warning
- app_monitor.c: removed no-op null assignment
- zenithd.c: sigaction() for SIGPIPE (was mixing signal/sigaction)

11 files changed

+++ native/lib/ipc_protocol.h
@@ -60,7 +60,7 @@ typedef struct __attribute__((packed)) {
 /* ---- Payloads ---- */
 
 typedef struct __attribute__((packed)) {
-    uint8_t  foreground_pid;
+    uint16_t foreground_pid;
     uint32_t current_profile_id;
     uint8_t  thermal_zone_count;
     struct {
